+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0 (DEV)
 
 - Add: Selection tool with basic Cut/Copy/Paste/Delete functionality.
+- Add: Button to locate primary Murphy in level.
 - Enh: Optimized rendering for level tiles. Now only visible range of tiles are
   rendering. The price is some flickering while scrolling.
 - Enh: Optimized editing. Now most operations with adequate level sizes runs